requirements.txt
streamlit
openai
pandas
plotly
python-docx
sentence-transformers
scikit-learn